# warnings: speed up work done in catch_warnings_for_item()