@bluetech 2020-05-07
> lovely, at first glance the osx failure seems unrelated

Thanks!

Yeah, the failure happens also without this PR, I am debugging it.